+++ Models/BankAccount/Abstract/IBankAccount.cs
@@ -2,11 +2,11 @@
 {
     public interface IBankAccount
     {
-        void SendMoney<T>(T toAccount, float amount) where T : BankAccountBase;
+        void SendMoney<T>(T toAccount, decimal amount) where T : BankAccountBase;
 
-        void GetMoney<T>(T fromAccount, float amount) where T : BankAccountBase;
+        void GetMoney<T>(T fromAccount, decimal amount) where T : BankAccountBase;
 
-        void Deposit(float amount);
+        void Deposit(decimal amount);
 
         void OpenAccount();
