+++ Models/BankAccount/Currencies/Dollars.cs
@@ -4,7 +4,7 @@ namespace BankManagment.Models.BankAccount.Currencies
 {
     public class Dollars : Currency
     {
-        public Dollars(float amount = 0) : base("Доллар", 840, amount)
+        public Dollars(decimal amount = 0) : base("Доллар", 840, amount)
         {
         }
     }
