+++ Models/BankAccount/Currencies/Rubles.cs
@@ -4,7 +4,7 @@ namespace BankManagment.Models.BankAccount.Currencies
 {
     public class Rubles : Currency
     {
-        public Rubles(float amount = 0) : base("Рубль", 810, amount)
+        public Rubles(decimal amount = 0) : base("Рубль", 810, amount)
         {
         }
     }
